prestamos desde articulos con buscador para socio no eliminar articulo PRESTADO

+++ app/src/main/java/db/ArticulosSQLite.kt
@@ -7,6 +7,7 @@ import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
 import android.util.Log
+import android.widget.Toast
 import pf.dam.articulos.EstadoArticulo
 import pf.dam.prestamos.EstadoPrestamo
 
@@ -30,6 +31,7 @@ class ArticulosSQLite (context: Context):
             )
         """.trimIndent()
         db?.execSQL(crearTablaArticulos)
+
     }
 
     override fun onUpgrade(db: SQLiteDatabase?, oldVersion: Int, newVersion: Int) {
@@ -193,12 +195,27 @@ class ArticulosSQLite (context: Context):
         db.close()
     }
 
-    fun borrarArticulo(idArticulo: Int): Int {
-        val db = writableDatabase
+//    fun borrarArticulo(idArticulo: Int): Int {
+//        val db = writableDatabase
+//        val rowsAffected = db.delete("articulos", "idArticulo = ?", arrayOf(idArticulo.toString()))
+//        db.close()
+//        return rowsAffected
+//    }
+fun borrarArticulo(idArticulo: Int, context: Context): Int {
+    val db = writableDatabase
+    val articulo = obtenerArticuloPorId(idArticulo) // Obtener el artículo por ID
+
+    if (articulo?.estado == EstadoArticulo.PRESTADO) {
+        // Mostrar un Toast si el estado es "prestado"
+        Toast.makeText(context, "No se puede borrar el artículo porque está prestado", Toast.LENGTH_SHORT).show()
+        return 0 // Indicar que no se borró el artículo
+    } else {
+        // Borrar el artículo si el estado no es "prestado"
         val rowsAffected = db.delete("articulos", "idArticulo = ?", arrayOf(idArticulo.toString()))
         db.close()
         return rowsAffected
     }
+}
 
     fun borrarTodosLosArticulos() {
         val db = writableDatabase
@@ -283,15 +300,15 @@ class ArticulosSQLite (context: Context):
             Log.d(TAG, "No se pudo actualizar el estado del artículo con ID: $idArticulo")
         }
     }
-   /* fun estaArticuloEnPrestamo(idArticulo: Int): Boolean {
+    fun articuloEnPrestamo(idArticulo: Int): Boolean {
         val db = readableDatabase
         val cursor = db.rawQuery(
             "SELECT * FROM prestamos WHERE idArticulo = ? AND estado = ?",
-            arrayOf(idArticulo.toString(), EstadoPrestamo.ACTIVO.toString())
+            arrayOf(idArticulo.toString(), EstadoPrestamo.ACTIVO.name)
         )
         val estaEnPrestamo = cursor.count > 0
         cursor.close()
         db.close()
         return estaEnPrestamo
-    }*/
+    }
 }
